Modify Firmware upload process for BiBoard
* Factory Reset Button:
After uploading the firmware, it will be forced into configuration mode. For questions, the user needs to manually select the yes or no button.
* Upgrade button:
After uploading the firmware, it is not forced to enter the configuration mode, but if it is in the configuration mode, it will automatically reply No to skip the prompt question.